requirements.txt
langchain-core
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
langchain-ollama
langchain-text-splitters
python-dotenv
chromadb
transformers
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
pypdf
python-docx
docx2txt
flask
flask-cors